PDMPFlux

Documentation for PDMPFlux.

PDMPFlux.AbstractPDMP

PDMPFlux.BoundBox

PDMPFlux.PDMPState

PDMPFlux.SpeedUpZigZag

PDMPFlux.StickyZigZag

PDMPFlux.ZigZag

PDMPFlux.ac_step

PDMPFlux.ac_step_with_proxy

PDMPFlux.erroneous_acceptance_rate

PDMPFlux.get_event_state

PDMPFlux.get_event_state

PDMPFlux.if_accept

PDMPFlux.if_reject

PDMPFlux.init_state

PDMPFlux.move_to_axes_and_stick

PDMPFlux.move_to_horizon

PDMPFlux.move_to_horizon2

PDMPFlux.moves_until_horizon

PDMPFlux.moves_until_horizon_or_axes

PDMPFlux.next_event

PDMPFlux.one_step_of_thinning

PDMPFlux.sample

PDMPFlux.sample_from_skeleton

PDMPFlux.sample_skeleton

PDMPFlux.sample_skeleton

PDMPFlux.set_AD_backend

PDMPFlux.thaw_one_coordinate

PDMPFlux.traj_for_animation

PDMPFlux.upper_bound_constant

PDMPFlux.upper_bound_grid

PDMPFlux.upper_bound_grid_vect

PDMPFlux.AbstractPDMP — Type

In PDMPFlux, every PDMP sampler is defined to be of subtype of AbstractPDMP.

PDMPFlux.BoundBox — Type

BoundBox <: Any

上界関数の出力を格納するための構造体．

Attributes:
    grid (Float64 Array{T,1}): The grid values.
    box_max (Float64 Array{T,1}): The maximum values on each segment of the grid.
    cum_sum (Float64 Array{T,1}): The cumulative sum of box_max.
    step_size (Float64): The step size of the grid.

PDMPFlux.PDMPState — Type

PDMPState <: Any

AbstractPDMP の状態空間の元．次のフィールドを持つ構造体として実装：

Attributes:
    x (Array{Float64, 1}): position
    v (Array{Float64, 1}): velocity
    t (Float64): time
    is_active (Array{Bool, 1}): indicator for the freezing state. Used in the sampling loop for Sticky samplers.

    ∇U (Function): gradient of the potential function
    rate (Function): rate function
    flow (Function): flow function
    velocity_jump (Function): velocity jump function
    upper_bound_func (Function): upper bound function
    upper_bound (Union{Nothing, NamedTuple}): upper bound box

    lambda_bar (Float64): upper bound for the Poisson process
    exp_rv (Float64): exponential random variable for the Poisson process
    
    lambda_t (Float64): rate at the current time
    horizon (Float64): horizon
    tp (Float64): time to the next event
    ts (Float64): time spent
    tt (Vector{Float64}): remaining frozen time for each coordinate, with dimension `dim`
    ar (Float64): acceptance rate for the thinning

    adaptive (Bool): adaptive indicator
    accept (Bool): accept indicator for the thinning
    stick_or_thaw_event (Bool): indicator for the sticking or thawing event
    
    errored_bound (Int): count of the number of errors in the upper bound
    rejected (Int): count of the number of rejections in the thinning
    hitting_horizon (Int): count of the number of hits of the horizon
    
    key (Any): random key

PDMPFlux.SpeedUpZigZag — Type

SpeedUpZigZag(dim::Int, ∇U::Function; grid_size::Int=10, tmax::Float64=1.0, 
    vectorized_bound::Bool=true, signed_bound::Bool=true, adaptive::Bool=true, kwargs...)

arguments for constructor

dim::Int: 空間の次元。

∇U::Function: ポテンシャルエネルギー関数の勾配。

grid_size::Int: 空間を離散化するためのグリッドポイントの数。デフォルトは10。

tmax::Float64: グリッドのホライズン。デフォルトは1.0。0の場合、適応的なtmaxが使用されます。

vectorized_bound::Bool: 境界にベクトル化された戦略を使用するかどうか。デフォルトはtrue。

signed_bound::Bool: 符号付き境界戦略を使用するかどうか。デフォルトはtrue。

adaptive::Bool: 適応的なtmaxを使用するかどうか。デフォルトはtrue。

kwargs...: その他のキーワード引数。

attributes of a ZigZag construct

dim::Int: 空間の次元。

refresh_rate::Float64: リフレッシュレート。

∇U::Function: ポテンシャルの勾配。

grid_size::Int: 空間を離散化するためのグリッドポイントの数。

tmax::Float64: グリッドのtmax。

adaptive::Bool: 適応的なtmaxを使用するかどうか。

vectorized_bound::Bool: ベクトル化された戦略を使用するかどうか。

signed_bound::Bool: 符号付き戦略を使用するかどうか。

flow::Function: インテグレータ関数。

rate::Array: プロセスのレート。

rate_vect::Array: ベクトル化されたレート。

signed_rate::Array: 符号付きレート。

signed_rate_vect::Array: ベクトル化され符号付きのレート。

velocity_jump::Function: 速度ジャンプ関数。

state: ZigZagサンプラーの状態。

PDMPFlux.StickyZigZag — Type

StickyZigZag(dim::Int, ∇U::Function; grid_size::Int=10, tmax::Float64=1.0, 
    vectorized_bound::Bool=true, signed_bound::Bool=true, adaptive::Bool=true, kwargs...)

arguments for constructor

dim::Int: dimension of the parameter space

∇U::Function: gradient of the potential function (= negative log-likelihood function)

κ::Vector{Float64}: prior inclusion probability. default is fill(0.5, dim).

grid_size::Int: number of the grid points for discretization of the parameter space. default is 10.

tmax::Float64: グリッドのホライズン．デフォルトは1.0．0の場合、適応的なtmaxが使用されます．

vectorized_bound::Bool: 境界にベクトル化された戦略を使用するかどうか．デフォルトはtrue．

signed_bound::Bool: 符号付き境界戦略を使用するかどうか．デフォルトはtrue．

adaptive::Bool: 適応的なtmaxを使用するかどうか．デフォルトはtrue．

kwargs...: その他のキーワード引数．

attributes of a ZigZag construct

dim::Int: 空間の次元．

refresh_rate::Float64: リフレッシュレート．

∇U::Function: ポテンシャルの勾配．

grid_size::Int: 空間を離散化するためのグリッドポイントの数．

tmax::Float64: グリッドのtmax．

adaptive::Bool: 適応的なtmaxを使用するかどうか．

vectorized_bound::Bool: ベクトル化された戦略を使用するかどうか．

signed_bound::Bool: 符号付き戦略を使用するかどうか．

flow::Function: インテグレータ関数．

rate::Array: プロセスのレート．

rate_vect::Array: ベクトル化されたレート．

signed_rate::Array: 符号付きレート．

signed_rate_vect::Array: ベクトル化され符号付きのレート．

velocity_jump::Function: 速度ジャンプ関数．

state: ZigZagサンプラーの状態．

PDMPFlux.ZigZag — Type

ZigZag(dim::Int, ∇U::Function; grid_size::Int=10, tmax::Float64=1.0, 
    vectorized_bound::Bool=true, signed_bound::Bool=true, adaptive::Bool=true, kwargs...)

arguments for constructor

dim::Int: 空間の次元。

∇U::Function: ポテンシャルエネルギー関数の勾配。

grid_size::Int: 空間を離散化するためのグリッドポイントの数。デフォルトは10。

tmax::Float64: グリッドのホライズン。デフォルトは1.0。0の場合、適応的なtmaxが使用されます。

vectorized_bound::Bool: 境界にベクトル化された戦略を使用するかどうか。デフォルトはtrue。

signed_bound::Bool: 符号付き境界戦略を使用するかどうか。デフォルトはtrue。

adaptive::Bool: 適応的なtmaxを使用するかどうか。デフォルトはtrue。

kwargs...: その他のキーワード引数。

attributes of a ZigZag construct

dim::Int: 空間の次元。

refresh_rate::Float64: リフレッシュレート。

∇U::Function: ポテンシャルの勾配。

grid_size::Int: 空間を離散化するためのグリッドポイントの数。

tmax::Float64: グリッドのtmax。

adaptive::Bool: 適応的なtmaxを使用するかどうか。

vectorized_bound::Bool: ベクトル化された戦略を使用するかどうか。

signed_bound::Bool: 符号付き戦略を使用するかどうか。

flow::Function: インテグレータ関数。

rate::Array: プロセスのレート。

rate_vect::Array: ベクトル化されたレート。

signed_rate::Array: 符号付きレート。

signed_rate_vect::Array: ベクトル化され符号付きのレート。

velocity_jump::Function: 速度ジャンプ関数。

state: ZigZagサンプラーの状態。

PDMPFlux.ac_step — Method

acceptance-rejection step
lambda_bar は正確な上界ではなく，grid が粗い場合に足りない可能性がある．
その場合は erroneous_acceptance_rate() で補正する．

PDMPFlux.ac_step_with_proxy — Method

代理上界 lambda_bar で足りた場合は，ここで簡単に Poisson 剪定を行う．
Poisson 剪定中に horizon を超えた場合は move_to_horizon2() を呼び出す．

PDMPFlux.erroneous_acceptance_rate — Method

代理上界 lambda_bar で足りなかった場合は horizon を縮めてより慎重に ac_step() を繰り返す．
state.adaptive = true の場合は，ここで horizon を恒久的に縮めておく．

PDMPFlux.get_event_state — Method

get_event_state(state::PDMPState, sampler::AbstractPDMP)::PDMPState

the argument `sampler` is only used to implement multiple dispatch.

state.accept が false である限り，one_step_of_thinning() を繰り返す．
state.accept が true になるには ac_step_with_proxy() → if_accept() が呼ばれる必要がある．

PDMPFlux.get_event_state — Method

multiple dispatch for get_event_state() function for StickyZigZag

Here, not only an accepted move, but also a crossing of axes and thawing a coordinate constitutes an event,
leading out of the loop.

When the new types of events are identified, it is informed via the `indicator` flag.

PDMPFlux.if_accept — Method

代理上界 lambda_bar を用いた剪定で accept された場合の処置
ここで one_step_of_thinning() を終了するために accept = true とされる．

PDMPFlux.if_reject — Method

代理上界 lambda_bar を用いた剪定で accept されなかった場合の処置
horizon を超えるまで Poisson 剪定を繰り返す．

PDMPFlux.init_state — Method

init_state():
PDMP オブジェクトの状態を初期化する．

Args:
    xinit (Float[Array, "dim"]): The initial position.
    vinit (Float[Array, "dim"]): The initial velocity.
    seed (int): The seed for random number generation.
    upper_bound_vect (bool, optional): Whether to use vectorized upper bound function. Defaults to False.
    signed_rate (bool, optional): Whether to use signed rate function. Defaults to False.
    adaptive (bool, optional): Whether to use adaptive upper bound. Defaults to False.
    constant_bound (bool, optional): Whether to use constant upper bound. Defaults to False.

Returns:
    PDMPState: The initialized PDMP state.

PDMPFlux.move_to_axes_and_stick — Method

movement special to `StickyPDMP`` samplers.
They move to one of the axes, and start to stick to it.

PDMPFlux.move_to_horizon — Method

tp > state.horizon の場合，もう一度 Poisson simulation を行う．

PDMPFlux.move_to_horizon2 — Method

代理上界 lambda_bar を使った Poisson 剪定中に horizon を超えた場合の動き

PDMPFlux.moves_until_horizon — Method

tp <= state.horizon の場合の処理
state.accept = true になるまでの ac_step() の繰り返しとして実装される．

PDMPFlux.moves_until_horizon_or_axes — Method

called while 
    1. no axis is crossed
    and
    2. min(tp, tt) <= state.horizon
This function calls ac_step() until
    1. state.accept = true
    or
    2. tp > state.horizon
    or
    3. an axis is crossed

PDMPFlux.next_event — Method

next_event(boundbox, exp_rv):
BoundBox オブジェクトを用いて次のイベント時間を Poisson 剪定によりシミュレーションする．
イベント時刻 t_prop とその直前の grid 点での上界の値を返す．

Args:
    boundbox: The boundbox object containing the cumulative sum and grid values.
    exp_rv: The exponential random variable.

Returns:
    A tuple containing the next event time (t_prop) and the corresponding upper bound value.

PDMPFlux.one_step_of_thinning — Method

state.accept が false である限り実行する処理
moves_until_horizon() → ac_step_with_proxy() → if_accept() が呼ばれるまで繰り返す．

PDMPFlux.sample — Method

sample()：PDMPSampler からサンプルをするための関数．
sample_skeleton() と sample_from_skeleton() の wrapper．

Args:
    N_sk (Int): Number of skeleton points to generate.
    N_samples (Int): Number of final samples to generate from the skeleton.
    xinit (Array{Float64, 1}): Initial position.
    vinit (Array{Float64, 1}): Initial velocity.
    seed (Int): Seed for random number generation.
    verbose (Bool, optional): Whether to print progress information. Defaults to true.

Returns:
    Array{Float64, 2}: Array of samples generated from the PDMP model.

PDMPFlux.sample_from_skeleton — Method

スケルトンからサンプリングをし，各行ベクトルに次元毎の時系列が格納された Matrix{Float64} を返す．

Args: N (Int): The number of samples to generate. output (PdmpOutput): The PDMP output containing the trajectory information.

Returns: Array{Float64, 2}: The sampled points from the PDMP trajectory skeleton.

PDMPFlux.sample_skeleton — Method

failsafe dispatch of sample_skeleton(), admitting scalar initial values, used mainly for 1d case.

PDMPFlux.sample_skeleton — Method

sample_skeleton(): PDMP Samplers からスケルトンを抽出する．

Parameters:
- n_sk (Int): The number of skeleton samples to generate.
- xinit (Array{Float64, 1}): The initial position of the particles.
- vinit (Array{Float64, 1}): The initial velocity of the particles.
- seed (Int): The seed value for random number generation.
- verbose (Bool): Whether to display progress bar during sampling. Default is true.

Returns:
- output: The output state of the sampling process.

PDMPFlux.set_AD_backend — Method

Set the AD backend for the <sampler name>AD() constructors defined in src/Samplers/*.jl

PDMPFlux.thaw_one_coordinate — Method

Now `tt` is reached, time to thaw one coordinate,
which is determined according to the proportion of `state.κ[state.is_active]`.

PDMPFlux.traj_for_animation — Method

traj_for_animation(): アニメーション用の軌道を抽出する．

Parameters:
- trajectory (Matrix{Float64}): The trajectory to be animated.
- N_start (Int): The starting index of the trajectory.
- N_max (Int): The ending index of the trajectory.
- coordinate_numbers (Array{Int, 1}): The indices of the coordinates to be plotted.
- dt (Float64): The time step for the animation.

Returns:
- traj: The snapshots of the trajectory to be animated.
- event_indeces: The indices of traj, where the events occur.

PDMPFlux.upper_bound_constant — Function

upper_bound_constant(func, a, b, n_grid=100, refresh_rate=0.0)
Computes the constant upper bound using the Brent's algorithm.
Brent のアルゴリズムを通じて定数でバウンドすることを試みる．必然的に n_grid=2．

Parameters:
- func: The function for which the upper bound constant is computed.
- a: The lower bound of the interval.
- b: The upper bound of the interval.
- n_grid: The number of grid points used for computation (default: 100).
- refresh_rate: The refresh rate for the upper bound constant (default: 0).

Returns:
- Tuple: A tuple containing the grid, box_max, cum_sum, and interval length.

PDMPFlux.upper_bound_grid — Function

upper_bound_grid(func, start=0.0, horizon, n_grid, refresh_rate)
Compute the upper bound as a piecewise constant function using a grid mechanism.

Args:
    func: the function for which the upper bound is computed
    start (Float64): the lower bound of the interval
    horizon (Float64): the upper bound of the interval
    n_grid (Int64, optional): size of the grid for the upperbound of func. Defaults to 100.
    refresh_rate (Float64, optional): refresh rate for the upper bound. Defaults to 0.

Returns:
    BoundBox: An object containing the upper bound constant information.

PDMPFlux.upper_bound_grid_vect — Function

upper_bound_grid_vect(func, start, horizon, n_grid)
Compute the upper bound using a grid with the vectorized strategy

For this function, func(x) takes vector values with the dimension `dim`.

Args:
    func: the function for which the upper bound is computed
    a (Float64): the lower bound of the interval
    b (Float64): the upper bound of the interval
    n_grid (Int64, optional): size of the grid for the upperbound of func. Defaults to 100.

Returns:
    BoundBox: An object containing the upper bound constant information.