PDMPFlux

Documentation for PDMPFlux.

PDMPFlux.AbstractPDMP

PDMPFlux.BoundBox

PDMPFlux.PDMPState

PDMPFlux.ZigZag

PDMPFlux.error_acceptance

PDMPFlux.if_accept

PDMPFlux.if_not_accept

PDMPFlux.init_state

PDMPFlux.inner_while

PDMPFlux.move_before_horizon

PDMPFlux.move_to_horizon

PDMPFlux.move_to_horizon2

PDMPFlux.next_event

PDMPFlux.ok_acceptance

PDMPFlux.one_step

PDMPFlux.one_step_while

PDMPFlux.sample

PDMPFlux.sample_from_skeleton

PDMPFlux.sample_skeleton

PDMPFlux.upper_bound_constant

PDMPFlux.upper_bound_grid

PDMPFlux.upper_bound_grid_vect

PDMPFlux.AbstractPDMP — Type

In PDMPFlux, every PDMP sampler is defined to be of subtype of AbstractPDMP.

PDMPFlux.BoundBox — Type

BoundBox <: Any

上界関数の出力を格納するための構造体．

Attributes:
    grid (Float64 Array{T,1}): The grid values.
    box_max (Float64 Array{T,1}): The maximum values on each segment of the grid.
    cum_sum (Float64 Array{T,1}): The cumulative sum of box_max.
    step_size (Float64): The step size of the grid.

PDMPFlux.PDMPState — Type

PDMPState <: Any

AbstractPDMP の状態空間の元．次のフィールドを持つ構造体として実装：

Attributes:
    x (Array{Float64, 1}): position
    v (Array{Float64, 1}): velocity
    t (Float64): time
    horizon (Float64): horizon
    key (Any): random key
    integrator (Function): integrator function
    grad_U (Function): gradient of the potential function
    rate (Function): rate function
    velocity_jump (Function): velocity jump function
    upper_bound_func (Function): upper bound function
    accept (Bool): accept indicator for the thinning
    upper_bound (Union{Nothing, NamedTuple}): upper bound box
    indicator (Bool): indicator for jumping
    tp (Float64): time to the next event
    ts (Float64): time spent
    exp_rv (Float64): exponential random variable for the Poisson process
    lambda_bar (Float64): upper bound for the Poisson process
    lambda_t (Float64): rate at the current time
    ar (Float64): acceptance rate for the thinning
    error_bound (Int): count of the number of errors in the upper bound
    rejected (Int): count of the number of rejections in the thinning
    hitting_horizon (Int): count of the number of hits of the horizon
    adaptive (Bool): adaptive indicator

PDMPFlux.ZigZag — Type

ZigZag(dim::Int, grad_U::Function; grid_size::Int=10, tmax::Float64=1.0, 
    vectorized_bound::Bool=true, signed_bound::Bool=true, adaptive::Bool=true, kwargs...)

ZigZag クラスは ZigZag サンプラーのためのクラスです。

引数

dim::Int: 空間の次元。

grad_U::Function: ポテンシャルエネルギー関数の勾配。

grid_size::Int: 空間を離散化するためのグリッドポイントの数。デフォルトは10。

tmax::Float64: グリッドのホライズン。デフォルトは1.0。0の場合、適応的なtmaxが使用されます。

vectorized_bound::Bool: 境界にベクトル化された戦略を使用するかどうか。デフォルトはtrue。

signed_bound::Bool: 符号付き境界戦略を使用するかどうか。デフォルトはtrue。

adaptive::Bool: 適応的なtmaxを使用するかどうか。デフォルトはtrue。

kwargs...: その他のキーワード引数。

属性

dim::Int: 空間の次元。

refresh_rate::Float64: リフレッシュレート。

grad_U::Function: ポテンシャルの勾配。

grid_size::Int: 空間を離散化するためのグリッドポイントの数。

tmax::Float64: グリッドのtmax。

adaptive::Bool: 適応的なtmaxを使用するかどうか。

vectorized_bound::Bool: ベクトル化された戦略を使用するかどうか。

signed_bound::Bool: 符号付き戦略を使用するかどうか。

integrator::Function: インテグレータ関数。

rate::Array: プロセスのレート。

rate_vect::Array: ベクトル化されたレート。

signed_rate::Array: 符号付きレート。

signed_rate_vect::Array: ベクトル化され符号付きのレート。

velocity_jump::Function: 速度ジャンプ関数。

state: ZigZagサンプラーの状態。

PDMPFlux.error_acceptance — Method

代理上界 lambda_bar で足りなかった場合は，ここで horizon を縮めて再度 Poisson 剪定を行う．

PDMPFlux.if_accept — Method

代理上界 lambda_bar を用いた剪定で accept された場合の処置
ここで one_step_while() を終了するために indicator = true とされる．

PDMPFlux.if_not_accept — Method

代理上界 lambda_bar を用いた剪定で accept されなかった場合の処置
horizon を超えるまで Poisson 剪定を繰り返す．

PDMPFlux.init_state — Method

init_state():
PDMP オブジェクトの状態を初期化する．

Args:
    xinit (Float[Array, "dim"]): The initial position.
    vinit (Float[Array, "dim"]): The initial velocity.
    seed (int): The seed for random number generation.
    upper_bound_vect (bool, optional): Whether to use vectorized upper bound function. Defaults to False.
    signed_rate (bool, optional): Whether to use signed rate function. Defaults to False.
    adaptive (bool, optional): Whether to use adaptive upper bound. Defaults to False.
    constant_bound (bool, optional): Whether to use constant upper bound. Defaults to False.

Returns:
    PdmpState: The initialized PDMP state.

PDMPFlux.inner_while — Method

lambda_bar は正確な上界ではなく，grid が粗い場合に足りない可能性がある．
その場合は error_acceptance() で補正する．

PDMPFlux.move_before_horizon — Method

horizon より先に event が起こった場合の処理
inner_while() の繰り返しとして実装される．

PDMPFlux.move_to_horizon — Method

event が horizon の先に起こった場合，もう一度 Poisson simulation を行う．

PDMPFlux.move_to_horizon2 — Method

代理上界 lambda_bar を使った Poisson 剪定中に horizon を超えた場合の動き

PDMPFlux.next_event — Method

next_event(boundbox, exp_rv):
BoundBox オブジェクトを用いて次のイベント時間を Poisson 剪定によりシミュレーションする．
イベント時刻 t_prop とその直前の grid 点での上界の値を返す．

Args:
    boundbox: The boundbox object containing the cumulative sum and grid values.
    exp_rv: The exponential random variable.

Returns:
    A tuple containing the next event time (t_prop) and the corresponding upper bound value.

PDMPFlux.ok_acceptance — Method

代理上界 lambda_bar で足りた場合は，ここで簡単に Poisson 剪定を行う．
Poisson 剪定中に horizon を超えた場合は move_to_horizon2() を呼び出す．

PDMPFlux.one_step — Method

one_step(state::PDMPState)::PDMPState

PDMPFlux.one_step_while — Method

state.indicator が false である限り実行する処理
move_before_horizon() で ok_acceptance() が呼ばれるまで繰り返す．

PDMPFlux.sample — Method

sample()：PDMPSampler からサンプルをするための関数．
sample_skeleton() と sample_from_skeleton() の wrapper．

Args:
    N_sk (Int): Number of skeleton points to generate.
    N_samples (Int): Number of final samples to generate from the skeleton.
    xinit (Array{Float64, 1}): Initial position.
    vinit (Array{Float64, 1}): Initial velocity.
    seed (Int): Seed for random number generation.
    verbose (Bool, optional): Whether to print progress information. Defaults to true.

Returns:
    Array{Float64, 2}: Array of samples generated from the PDMP model.

PDMPFlux.sample_from_skeleton — Method

スケルトンからサンプリングをし，各行ベクトルに次元毎の時系列が格納された Matrix{Float64} を返す．

Args: N (Int): The number of samples to generate. output (PdmpOutput): The PDMP output containing the trajectory information.

Returns: Array{Float64, 2}: The sampled points from the PDMP trajectory skeleton.

PDMPFlux.sample_skeleton — Method

sample_skeleton(): PDMP Samplers からスケルトンを抽出する．

Parameters:
- n_sk (Int): The number of skeleton samples to generate.
- xinit (Array{Float64, 1}): The initial position of the particles.
- vinit (Array{Float64, 1}): The initial velocity of the particles.
- seed (Int): The seed value for random number generation.
- verbose (Bool): Whether to display progress bar during sampling. Default is true.

Returns:
- output: The output state of the sampling process.

PDMPFlux.upper_bound_constant — Function

upper_bound_constant(func, a, b, n_grid=100, refresh_rate=0.0)
Computes the constant upper bound using the Brent's algorithm.
Brent のアルゴリズムを通じて定数でバウンドすることを試みる．必然的に n_grid=2．

Parameters:
- func: The function for which the upper bound constant is computed.
- a: The lower bound of the interval.
- b: The upper bound of the interval.
- n_grid: The number of grid points used for computation (default: 100).
- refresh_rate: The refresh rate for the upper bound constant (default: 0).

Returns:
- Tuple: A tuple containing the grid, box_max, cum_sum, and interval length.

PDMPFlux.upper_bound_grid — Function

upper_bound_grid(func, start=0.0, horizon, n_grid, refresh_rate)
Compute the upper bound as a piecewise constant function using a grid mechanism.

Args:
    func: the function for which the upper bound is computed
    start (Float64): the lower bound of the interval
    horizon (Float64): the upper bound of the interval
    n_grid (Int64, optional): size of the grid for the upperbound of func. Defaults to 100.
    refresh_rate (Float64, optional): refresh rate for the upper bound. Defaults to 0.

Returns:
    BoundBox: An object containing the upper bound constant information.

PDMPFlux.upper_bound_grid_vect — Function

upper_bound_grid_vect(func, start, horizon, n_grid)
Compute the upper bound using a grid with the vectorized strategy

For this function, func(x) takes vector values with the dimension `dim`.

Args:
    func: the function for which the upper bound is computed
    a (Float64): the lower bound of the interval
    b (Float64): the upper bound of the interval
    n_grid (Int64, optional): size of the grid for the upperbound of func. Defaults to 100.

Returns:
    BoundBox: An object containing the upper bound constant information.